feat: migrate to TypeScript api folder
